+++ src/components/Carousel.tsx
@@ -1,3 +1,4 @@
+import { PostImage } from "@/shared/interfaces";
 import { wrap } from "@popmotion/popcorn";
 import { AnimatePresence, motion } from "framer-motion";
 import { useState } from "react";
@@ -7,7 +8,7 @@ import {
 } from "react-icons/io";
 
 interface Props {
-  images: string[];
+  images: PostImage[];
 }
 
 const Carousel = ({ images }: Props) => {
@@ -42,38 +43,42 @@ const Carousel = ({ images }: Props) => {
   };
 
   return (
-    <div className="carousel-wrapper relative flex justify-center items-center rounded-xl overflow-hidden">
-      <AnimatePresence initial={false} custom={direction}>
-        <motion.img
-          key={page}
-          src={images[imageIndex]}
-          custom={direction}
-          variants={variants}
-          initial="enter"
-          animate="center"
-          exit="exit"
-          transition={{
-            x: { type: "spring", stiffness: 300, damping: 30 },
-            opacity: { duration: 0.2 },
-          }}
-          drag="x"
-          dragConstraints={{ left: 0, right: 0 }}
-          dragElastic={1}
-          onDragEnd={(_, { offset, velocity }) => {
-            const swipe = swipePower(offset.x, velocity.x);
-
-            if (swipe < -swipeConfidenceThreshold) pagination(1);
-            else if (swipe > swipeConfidenceThreshold) pagination(-1);
-          }}
-        />
-      </AnimatePresence>
-      <div className="prev" onClick={() => pagination(1)}>
-        <IoMdArrowDropleftCircle />
+    <>
+      <div className="carousel-wrapper relative flex justify-center items-center rounded-xl overflow-hidden">
+        <AnimatePresence initial={false} custom={direction}>
+          <motion.img
+            key={page}
+            src={images[imageIndex].img}
+            custom={direction}
+            variants={variants}
+            initial="enter"
+            animate="center"
+            exit="exit"
+            transition={{
+              x: { type: "spring", stiffness: 300, damping: 30 },
+              opacity: { duration: 0.2 },
+            }}
+            drag="x"
+            dragConstraints={{ left: 0, right: 0 }}
+            dragElastic={1}
+            onDragEnd={(_, { offset, velocity }) => {
+              const swipe = swipePower(offset.x, velocity.x);
+              if (swipe < -swipeConfidenceThreshold) pagination(1);
+              else if (swipe > swipeConfidenceThreshold) pagination(-1);
+            }}
+          />
+        </AnimatePresence>
+        <div className="prev" onClick={() => pagination(-1)}>
+          <IoMdArrowDropleftCircle />
+        </div>
+        <div className="next" onClick={() => pagination(1)}>
+          <IoMdArrowDroprightCircle />
+        </div>
       </div>
-      <div className="next" onClick={() => pagination(-1)}>
-        <IoMdArrowDroprightCircle />
-      </div>
-    </div>
+      <p className="text-sm text-gray-300 text-center">
+        {images[imageIndex].caption}&nbsp;
+      </p>
+    </>
   );
 };
