+++ src/shared/interfaces/index.ts
@@ -1,0 +1,29 @@
+export interface PostImage {
+  img: string;
+  caption?: string;
+}
+
+export interface Post {
+  title: string;
+  status: "IC" | "GB";
+  tags: Tag[];
+  ic_link: string;
+  author: string;
+  author_pfp: string;
+  date: string;
+  description: string;
+  images: PostImage[];
+  comments: Comment[];
+}
+
+interface Tag {
+  name: string;
+  color: string;
+}
+
+interface Comment {
+  author: string;
+  author_pfp: string;
+  content: string;
+  date: string;
+}
